+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 1.3.0. (2023-06-26)
+
+* Logger is configured automatically for Logback and Log4j2 - no need to add any configuration to your application (if you have configured the OpenTelemetry logger already, it will be used)
+
 ## Version 1.2.0. (2023-06-06)
 
 * Set the base time unit to "seconds" - which ensures future compatibility with upcoming versions of the Grafana Agent